+++ src/test/java/com/odiga/fiesta/ControllerTestSupport.java
@@ -8,7 +8,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.odiga.fiesta.festival.controller.FestivalController;
+import com.odiga.fiesta.festival.controller.FestivalStaticDataController;
 import com.odiga.fiesta.festival.service.CategoryService;
 import com.odiga.fiesta.festival.service.CompanionService;
 import com.odiga.fiesta.festival.service.MoodService;
@@ -20,7 +20,7 @@ import com.odiga.fiesta.log.service.LogService;
 @WithMockUser
 @WebMvcTest(controllers = {
 	// 사용하는 컨트롤러 여기에 추가
-	FestivalController.class,
+	FestivalStaticDataController.class,
 	LogController.class
 })
 public abstract class ControllerTestSupport {
